Simplify private list delivery bindings
