+++ Package.swift
@@ -3,7 +3,7 @@
 import PackageDescription
 
 let tag = "v0.1.0-rc12"
-let checksum = "96e835e46e2342e99a2fb016e36909759c7f14568a6158580c1adc2ae9588dbd"
+let checksum = "7004689a42c3054e6d99156157a984afbac72446ed2c0d85208557cdaf6ef987"
 let url = "https://github.com/pubky/paykit-rs/releases/download/\(tag)/Paykit.xcframework.zip"
 
 let package = Package(
