+++ paykit-ffi/src/lib.rs
@@ -7,6 +7,7 @@ mod payment_adapter;
 mod payment_requests;
 mod payment_resolution;
 mod private_links;
+mod private_lists;
 mod profiles;
 mod receipts;
 mod sdk;
@@ -24,6 +25,7 @@ pub use payment_adapter::*;
 pub use payment_requests::*;
 pub use payment_resolution::*;
 pub use private_links::*;
+pub use private_lists::*;
 pub use profiles::*;
 pub use receipts::*;
 pub use sdk::*;
